requirements.txt
requests
pandas
numpy
matplotlib
plotly
statsmodels
scipy
python-dotenv
